# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application (PPA V5) API
  version: 1.0.0
  description: |
    RESTful API for a campus placement management system.
    Supports three roles: Admin, Company, and Student.
servers:
  - url: http://localhost:5001
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [admin, company, student]
        is_active:
          type: boolean
        is_blacklisted:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        message:
          type: string
        channel:
          type: string
        is_sent:
          type: boolean
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
    AsyncJob:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        job_type:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        file_path:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
                  profile:
                    type: object
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user info
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  profile:
                    type: object

  # ─── Notifications ────────────────────────────────────
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of notifications (max 50)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /api/notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark a notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Async Jobs ───────────────────────────────────────
  /api/jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get async job status
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncJob"

  # ─── Admin ────────────────────────────────────────────
...